Add information about option codes corresponding to exterior colors (#104)
I went to the Tesla design studio web site and matched up the exterior colors I saw on the site with the option codes here. The web site appears to use a "compositor" API endpoint to retrieve images for the designer based on color (and other options) chosen. I went to the web site and tracked updates to the image href to figure out which of the options changed when I switch colors. I hope this proves useful!