Create overwritable method called `base_uri`
We are making a system where we sometimes want to proxy requests to another server instead of the default BASE_URI. Given the existing code it is hard to do this. With this PR I make it easier to dynamically overwrite the base_uri with a subclass.

Here is an example that changes the base uri dynamically:

```ruby
class TeslaAdapter < TeslaApi::Client
  def base_uri
    "http://localhost/mock"
  end
end
```